pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.107.1",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "streamlit>=1.49.1",
]
